Add model and migration to DB (Usuarios)

+++ database/migrations/2023_01_16_155512_create_usuarios_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('nombre');
             $table->string('fotos');
-            $table->tinyInteger('estado');
+            $table->integer('estado');
             $table->timestamps();
         });
     }
